Introduced standardjs
A git hook was added to ensure that the conventions are respected.
The work was mostly done by the --fix option of the 'standard' tool.
Some linter exceptions were added to ensure minimal changes to the code:
- Buffer.from should not be used anymore: https://github.com/standard/standard/issues/693
- Errors were not handled in some cases
- Global variables: log, config ...
- callback(true) is not recommended: https://github.com/standard/standard/issues/623

Other than that '==' was used instead of '===' and some variables were redeclared in the same scope.